registry is a flow prop now.
