+++ apps/graph-editor/src/hooks/useRegistry.ts
