+++ packages/flow/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@behave-graph/core": "*",
-    "@behave-graph/scene": "*",
     "@fortawesome/fontawesome-svg-core": "^6.1.2",
     "@fortawesome/free-solid-svg-icons": "^6.1.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
