+++ packages/flow/src/index.ts
@@ -20,7 +20,6 @@ export * from './hooks/useFlowHandlers.js';
 export * from './hooks/useGraphRunner.js';
 export * from './hooks/useBehaveGraphFlow.js';
 export * from './hooks/useNodeSpecJson.js';
-export * from './hooks/useRegistry.js';
 export * from './hooks/useCustomNodeTypes.js';
 export * from './hooks/useMergeMap.js';
