--- packages/flow/src/util/getNodeSpecJSON.ts
@@ -1,7 +1,0 @@
-import { memo, NodeSpecJSON, writeNodeSpecsToJSON } from '@behave-graph/core';
-
-import { useRegistry } from '../hooks/useRegistry.js';
-
-export const getNodeSpecJSON = memo<NodeSpecJSON[]>(() =>
-  writeNodeSpecsToJSON(useRegistry())
-);
